requirements.txt
streamlit
python-docx
nltk